+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Fixed an issue that caused the like button not to update.
 - Added `fitBounds` method to map.
 - Fixed map content area display when multiple components are present.
+- Use infinite scroll to limit the number of items rendered in a feature list.
 
 # [0.3.1] - 2018-06-08
 - Prevent form submission if there are no fields in the page.